// src/interfaces/exchanges.interface.ts
export interface ExchangesRootState {
  exchanges: Exchange[];
  isLoading: boolean;
  isError: boolean;
}

export interface ExchangeRootState {
  exchange: FullExchange | null;
  isLoading: boolean;
  isError: boolean;
}

export interface Exchange {
  id: string;
  name: string;
  year_established: number;
  country: string;
  description: string;
  url: string;
  image: string;
  has_trading_incentive: boolean;
  trust_score: number;
  trust_score_rank: number;
  trade_volume_24h_btc: number;
  trade_volume_24h_btc_normalized: number;
}

export interface FullExchange {
  name: Name;
  year_established: number;
  country: string;
  description: string;
  url: string;
  image: string;
  facebook_url: string;
  reddit_url: string;
  telegram_url: string;
  slack_url: string;
  other_url_1: string;
  other_url_2: string;
  twitter_handle: TwitterHandle;
  has_trading_incentive: boolean;
  centralized: boolean;
  public_notice: string;
  alert_notice: string;
  trust_score: number;
  trust_score_rank: number;
  trade_volume_24h_btc: number;
  trade_volume_24h_btc_normalized: number;
  tickers: Ticker[];
  status_updates: StatusUpdate[];
}

export enum Name {
  Binance = 'Binance',
}

export interface StatusUpdate {
  description: string;
  category: string;
  created_at: string;
  user: string;
  user_title: string;
  pin: boolean;
  project: Project;
}

export interface Project {
  type: string;
  id: TwitterHandle;
  name: Name;
  image: Image;
}

export enum TwitterHandle {
  Binance = 'binance',
}

export interface Image {
  thumb: string;
  small: string;
  large: string;
}

export interface Ticker {
  base: string;
  target: string;
  market: Market;
  last: number;
  volume: number;
  converted_last: { [key: string]: number };
  converted_volume: { [key: string]: number };
  trust_score: string;
  bid_ask_spread_percentage: number;
  timestamp: string;
  last_traded_at: string;
  last_fetch_at: string;
  is_anomaly: boolean;
  is_stale: boolean;
  trade_url: string;
  token_info_url: null;
  coin_id: string;
  target_coin_id?: string;
}

export interface Market {
  name: Name;
  identifier: TwitterHandle;
  has_trading_incentive: boolean;
}